[Variations] Moving base class to utils
