+++ booking.py
@@ -7,6 +7,7 @@ from .utils import Weight
 from .utils import Action
 from .utils import Count
 from .utils import Histogram
+from .utils import Variation
 
 from ROOT import gROOT
 gROOT.SetBatch(True)
