+++ utils/__init__.py
@@ -14,3 +14,5 @@ from ._run import RDataFrameCutWeight
 
 from ._printing import Node as PrintedNode
 from ._printing import drawTree2
+
+from ._variations import Variation
